feat: use library for published tab
