+++ src/locales/ar/common.json
@@ -57,7 +57,6 @@
   "PAGE_TITLES": {
     "MY_GRAASP": "العناصر الخاصّة بي",
     "BOOKMARKED_ITEMS": "العناصر المفضلة",
-    "PUBLISHED_ITEMS": "العناصر المنشورة",
     "TRASH": "القمامة"
   }
 }
