+++ src/locales/ar/home.json
@@ -2,7 +2,6 @@
   "MENU": {
     "MY_ITEMS": "قبضتي",
     "BOOKMARKED_ITEMS": "العلامات المرجعية",
-    "PUBLISHED_ITEMS": "نُشرت",
     "RECYCLED_ITEMS": "مُعاد تدويره"
   }
 }
