+++ src/locales/de/common.json
@@ -57,7 +57,6 @@
   "PAGE_TITLES": {
     "MY_GRAASP": "Mein Graasp",
     "BOOKMARKED_ITEMS": "Lieblingselemente",
-    "PUBLISHED_ITEMS": "Veröffentlichte Elemente",
     "TRASH": "Papierkorb"
   }
 }
