+++ src/locales/de/home.json
@@ -2,7 +2,6 @@
   "MENU": {
     "MY_ITEMS": "Mein Graasp",
     "BOOKMARKED_ITEMS": "Merkliste",
-    "PUBLISHED_ITEMS": "Veröffentlicht",
     "RECYCLED_ITEMS": "Recycelt"
   }
 }
