+++ src/locales/en/common.json
@@ -57,7 +57,7 @@
   "PAGE_TITLES": {
     "MY_GRAASP": "My Graasp",
     "BOOKMARKED_ITEMS": "Bookmarks",
-    "PUBLISHED_ITEMS": "Published Items",
+    "LIBRARY": "Library",
     "TRASH": "Trash"
   },
   "USER_SWITCH": {
