+++ src/locales/en/home.json
@@ -3,11 +3,18 @@
     "HOME": "Home",
     "MY_ITEMS": "My Graasp",
     "BOOKMARKED_ITEMS": "Bookmarks",
-    "PUBLISHED_ITEMS": "Published",
+    "LIBRARY": "Library",
     "RECYCLED_ITEMS": "Trash"
   },
   "BOOKMARKED_ITEMS": {
     "SHOW_LESS": "Show less",
     "SHOW_MORE": "Show more"
+  },
+  "LIBRARY": {
+    "CARD": {
+      "TITLE": "Graasp Library",
+      "SUBTITLE": "Share and Discover Open Educational Resources in the Library",
+      "BUTTON_TEXT": "Explore"
+    }
   }
 }
