+++ src/locales/es/common.json
@@ -57,7 +57,6 @@
   "PAGE_TITLES": {
     "MY_GRAASP": "Mi Graasp",
     "BOOKMARKED_ITEMS": "Artículos favoritos",
-    "PUBLISHED_ITEMS": "Artículos publicados",
     "TRASH": "Basura"
   }
 }
