+++ src/locales/fr/common.json
@@ -57,7 +57,7 @@
   "PAGE_TITLES": {
     "MY_GRAASP": "Accueil",
     "BOOKMARKED_ITEMS": "Signets",
-    "PUBLISHED_ITEMS": "Publications",
+    "LIBRARY": "Bibliothèque",
     "TRASH": "Poubelle"
   },
   "USER_SWITCH": {
