+++ src/locales/it/common.json
@@ -56,7 +56,6 @@
   "UNKNOWN_DATE": "Sconosciuto",
   "PAGE_TITLES": {
     "MY_GRAASP": "Il mio Graasp",
-    "PUBLISHED_ITEMS": "Elementi pubblicati",
     "TRASH": "Cestino"
   }
 }
