+++ src/locales/it/home.json
@@ -2,7 +2,6 @@
   "MENU": {
     "MY_ITEMS": "Il mio Graasp",
     "BOOKMARKED_ITEMS": "Segnalibri",
-    "PUBLISHED_ITEMS": "Pubblicato",
     "RECYCLED_ITEMS": "Riciclato"
   }
 }
